+++ spec/graphql/static_validation/rules/directives_are_defined_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe GraphQL::StaticValidation::DirectivesAreDefined do
-  let(:document) { GraphQL.parse("
+  let(:query_string) {"
     query getCheese {
       okCheese: cheese(id: 1) {
         id @skip(if: true),
@@ -11,10 +11,11 @@ describe GraphQL::StaticValidation::DirectivesAreDefined do
         }
       }
     }
-  ")}
+  "}
 
   let(:validator) { GraphQL::StaticValidation::Validator.new(schema: DummySchema, rules: [GraphQL::StaticValidation::DirectivesAreDefined]) }
-  let(:errors) { validator.validate(document) }
+  let(:query) { GraphQL::Query.new(DummySchema, query_string) }
+  let(:errors) { validator.validate(query) }
 
   describe 'non-existent directives' do
     it 'makes errors for them' do