+++ spec/graphql/static_validation/rules/fragments_are_used_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe GraphQL::StaticValidation::FragmentsAreUsed do
-  let(:document) { GraphQL.parse("
+  let(:query_string) {"
     query getCheese {
       name,
       ...cheeseFields
@@ -9,10 +9,11 @@ describe GraphQL::StaticValidation::FragmentsAreUsed do
     }
     fragment cheeseFields on Cheese { fatContent }
     fragment unusedFields on Cheese { is, not, used }
-  ")}
+  "}
 
   let(:validator) { GraphQL::StaticValidation::Validator.new(schema: DummySchema, rules: [GraphQL::StaticValidation::FragmentsAreUsed]) }
-  let(:errors) { validator.validate(document) }
+  let(:query) { GraphQL::Query.new(DummySchema, query_string) }
+  let(:errors) { validator.validate(query) }
 
   it 'adds errors for unused fragment definitions' do
     assert_includes(errors, {"message"=>"Fragment unusedFields was defined, but not used", "locations"=>[{"line"=>8, "column"=>5}]})