+++ spec/graphql/static_validation/rules/variables_are_input_types_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe GraphQL::StaticValidation::VariablesAreInputTypes do
-  let(:document) { GraphQL.parse('
+  let(:query_string) {'
     query getCheese(
       $id:        Int = 1,
       $str:       [String!],
@@ -11,10 +11,11 @@ describe GraphQL::StaticValidation::VariablesAreInputTypes do
     ) {
       cheese(id: $id) { source }
     }
-  ')}
+  '}
 
   let(:validator) { GraphQL::StaticValidation::Validator.new(schema: DummySchema, rules: [GraphQL::StaticValidation::VariablesAreInputTypes]) }
-  let(:errors) { validator.validate(document) }
+  let(:query) { GraphQL::Query.new(DummySchema, query_string) }
+  let(:errors) { validator.validate(query) }
 
   it "finds variables whose types are invalid" do
     expected = [